requirements.txt
discord.py
python-dotenv
jokeapi
Flask>=2.0.0